Now fetches repo info from github
