+++ components/content/index.ts
@@ -1,2 +1,3 @@
 export * from './Hero';
 export * from './Projects';
+export * from './OpenSource';
