+++ components/ui/index.ts
@@ -1,4 +1,5 @@
 export * from './Button';
 export * from './Emoji';
+export * from './Github_Card';
 export * from './Project_Card';
 export * from './Section';
